Revert "split nodes into their own files"
This reverts commit 1446ebf9b56639eeebddc9f6df0ee570df73d68e.
