--- lib/rbexy/nodes/base.rb
@@ -1,9 +1,0 @@
-module Rbexy
-  module Nodes
-    class Base
-      def precompile
-        [self]
-      end
-    end
-  end
-end
