--- lib/rbexy/nodes/silent_newline.rb
@@ -1,9 +1,0 @@
-module Rbexy
-  module Nodes
-    class SilentNewline < Base
-      def compile
-        "\n"
-      end
-    end
-  end
-end
