--- lib/rbexy/nodes/util.rb
@@ -1,14 +1,0 @@
-module Rbexy
-  module Nodes
-    # TODO: will we still need this?
-    module Util
-      def self.safe_string(str)
-        str.gsub('"', '\\"')
-      end
-
-      def self.safe_tag_name(name)
-        name.gsub(".", "__")
-      end
-    end
-  end
-end
